+++ tests/shims/WebServer.h
@@ -1,4 +1,9 @@
 #pragma once
+#define UPLOAD_FILE_START 0
+#define UPLOAD_FILE_WRITE 1
+#define UPLOAD_FILE_END 2
+#define UPLOAD_FILE_ABORTED 3
+struct HTTPUpload { int status=0; unsigned currentSize=0; unsigned char* buf=nullptr; };
 #include <Arduino.h>
 enum HTTPMethod { HTTP_ANY, HTTP_GET, HTTP_POST, HTTP_OPTIONS, HTTP_DELETE };
 class WebServer {
@@ -6,6 +11,9 @@ public:
   WebServer(int){}
   void on(const char*, void(*)()){}
   void on(const char*, HTTPMethod, void(*)()){}
+  void on(const char*, HTTPMethod, void(*)(), void(*)()){}
+  HTTPUpload& upload(){ static HTTPUpload u; return u; }
+  void streamFile(class File&, const char*){}
   void onNotFound(void(*)()){}
   void begin(){}
   void handleClient(){}